+++ .gitignore
@@ -2,3 +2,5 @@
 GoogleMapsTestApp.xcodeproj/project.xcworkspace/xcuserdata/nikitamosencev.xcuserdatad/UserInterfaceState.xcuserstate
 
 GoogleMapsTestApp.xcworkspace/xcuserdata/nikitamosencev.xcuserdatad/UserInterfaceState.xcuserstate
+
+GoogleMapsTestApp.xcworkspace/xcuserdata/nikitamosencev.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
