--- GoogleMapsTestApp.xcworkspace/xcuserdata/nikitamosencev.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "74E8576C-7203-4483-A705-6211FA2F4A96"
-   type = "0"
-   version = "2.0">
-</Bucket>
