+++ GoogleMapsTestApp/UserStories/CustomViews/CustomProgressViewStyle.swift
@@ -34,22 +34,16 @@ struct BarProgressStyle: ProgressViewStyle {
                                 RoundedRectangle(cornerRadius: 10.0)
                                     .fill(color)
                                     .frame(width: geometry.size.width * progress, height: height)
-//                                VStack {
-//                                    Text("70 км/ч")
-//                                        .font(.system(size: 11))
-//                                        .foregroundColor(.gray)
-//                                        .padding(.bottom, 8)
-                                    
-                                    Circle()
-                                        .fill(color)
-                                        .shadow(color: .black.opacity(0.08), radius: 3, x: 0, y: 2)
-                                        .frame(width: 20, height: 20)
-                                        .overlay(
-                                            Circle()
-                                                .stroke(Color.white, lineWidth: 2)
-                                                .frame(width: 20, height: 20)
-                                        )
-//                                }
+
+                                Circle()
+                                    .fill(color)
+                                    .shadow(color: .black.opacity(0.08), radius: 3, x: 0, y: 2)
+                                    .frame(width: 20, height: 20)
+                                    .overlay(
+                                        Circle()
+                                            .stroke(Color.white, lineWidth: 2)
+                                            .frame(width: 20, height: 20)
+                                    )
                             }
                         }
                             .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .leading)
@@ -63,67 +57,3 @@ struct BarProgressStyle: ProgressViewStyle {
         }
     }
 }
-
-
-/*
- Text("70 км/ч")
-     .font(.system(size: 11))
-     .foregroundColor(.gray)
-     .padding(.bottom, 8)
- */
-
-
-//struct BarProgressStyle: ProgressViewStyle {
-////    var color: Color = .purple
-//    var height: Double = 20.0
-//    var labelFontStyle: Font = .body
-//
-//    func makeBody(configuration: Configuration) -> some View {
-//
-////        let progress = configuration.fractionCompleted ?? 0.0
-//
-////        GeometryReader { geometry in
-//
-////            VStack(alignment: .leading) {
-////                configuration.
-//
-////                configuration.label
-////                    .font(labelFontStyle)
-////
-////                RoundedRectangle(cornerRadius: 10.0)
-//////                    .fill(Color(uiColor: .systemGray5))
-////                    .frame(height: height)
-////                    .frame(width: geometry.size.width)
-////                    .overlay() {
-////                        RoundedRectangle(cornerRadius: 10.0)
-//////                            .fill(color)
-////                            .frame(width: geometry.size.width * progress)
-////                            .overlay (
-////                                if let currentValueLabel = configuration.currentValueLabel {
-////
-////                                    currentValueLabel
-////                                        .font(.headline)
-////                                        .foregroundColor(.white)
-////                                }
-////                            )
-////                    }
-//
-////            }
-
-//        }
-//    }
-//}
-
-//
-//VStack(alignment: .leading) {
-//    Circle()
-//        .fill(color)
-//        .frame(width: 10, height: 10)
-//        .border(.white)
-//        .overlay(
-//            Circle()
-//                .stroke(Color.white)
-//                .frame(width: 10, height: 10)
-//        )
-////                                .frame(width: geometry.size.width * progress)
-//}
